Move contracts to entities namespace
